Improve memory performance of redirects
For sites that have a large number of redirects, the existing
implementation will throw a `RangeError: Maximum call stack size
exceeded` error, as all the redirect jobs are first built into memory,
then spread out into the uploadQueue. This change tweaks the flow to
push the redirect jobs into the queue using a forEach loop instead,
which doesn't trigger this same error with a large number of redirects.